data-simulator/src/fan_controller/requirements.txt
paho-mqtt
requests
ssl
